+++ webapp/views/object_type_metadata.py
@@ -1,3 +1,5 @@
+from urllib.parse import quote
+
 from views.utils import (format_amr, format_cog, format_ko, format_orthogroup,
                          format_pfam, missing_mandatory, optional2status)
 
@@ -126,6 +128,19 @@ class PathwayMetadata(BaseObjectMetadata):
         return None
 
 
+class GroupMetadata(BaseObjectMetadata):
+
+    object_type = "group"
+    object_name = "Genome Group"
+    overview_description = "Overview of defined groups of genomes."
+
+    @staticmethod
+    def format_entry(entry, to_url=False):
+        if to_url:
+            return f"<a href=/groups/{quote(entry)}>{entry}</a>"
+        return entry
+
+
 class MetadataGetter():
 
     metadata_classes = [AmrMetadata, CogMetadata, KoMetadata, PfamMetadata,
@@ -148,6 +163,10 @@ class MetadataGetter():
     def get_orthology_metadata(self):
         return (self.object_type_to_metadata[object_type]
                 for object_type in self._orthology)
+
+    @property
+    def group_metadata(self):
+        return GroupMetadata()
 
 
 def my_locals(local_dico):